[SIDE-479] More refactoring
