+++ src/commonHostClient.ts
@@ -1,0 +1,8 @@
+import { ConnInfo } from './proto/broker_pb';
+
+/**
+ * @internal
+ */
+export interface CommonHostClient {
+  connect(conn: ConnInfo.AsObject): Promise<void>;
+}
